Fixed Facebook login bug
This bug was introduced when I added the WordPress blog post creation feature.

The prior version used to read my Facebook username and password from the `.env` file into the constants `USERNAME` and `PASSWORD`. With the addition of WordPress blog post creation, I updated the code to read the Facebook credentials into the constants `FACEBOOK_USERNAME` and `FACEBOOK_PASSWORD`, but forgot to update the Facebook login code to use the new names for those constants.